# proj_trans(): fix performance regression