@rouault 2022-12-11
Actually I was wrong. 9.1 is not affected. master only